% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_tss.R
\name{cn_tss}
\alias{cn_tss}
\title{Total sum of squares}
\usage{
cn_tss(y)
}
\arguments{
\item{y}{\code{numeric} vector on which to calculate the sum of squares.}
}
\value{
Return the sum of squares (\code{numerical}).
}
\description{
Compute the total sum of squares.
}
\details{
Computed through matrix multiplication.
}
\examples{
y <- rnorm(10)
cn_tss(y)

}
