% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_crit.R
\name{group_crit}
\alias{group_crit}
\title{Critical values}
\usage{
group_crit(x, division)
}
\arguments{
\item{x}{\code{numeric}, an ordered vector of elements.}

\item{division}{\code{matrix} of position that define a new group (rows) for different partitions (columns), as generated by the function \code{group_division}.}
}
\value{
Return a \code{matrix} of the same dimension as \code{division}, containing the average values of \code{x} for the position defined in the \code{division} and the position preceeding it.
}
\description{
Generate critical values associated with divisions.
}
\examples{
#General example
x <- sample(1:100, 10)
division <- group_division(x, n_group = 3, min_group = 2)
group_crit(x, division)
}
