% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_division_bound.R
\name{group_division_bound}
\alias{group_division_bound}
\title{Group division bound}
\usage{
group_division_bound(x, division, min_crit = NULL, max_crit = NULL)
}
\arguments{
\item{x}{\code{numeric}, an ordered vector of elements.}

\item{division}{\code{matrix} of position that define a new group (rows) for different partitions (columns), as generated by the function group_division; alternatively a \code{numeric} vector representing a single partition.}

\item{min_crit, max_crit}{\code{numeric} vectors of length corresponding to the number of group (number of rows of \code{division}) minus one, representing inequality constraints on each critical values in \code{x} \code{(min <= x <= max)}. The default value is \code{NULL} and represent no constraint.}
}
\value{
Return a \code{logical} vector indication which column of \code{division} that respect the constraints.
}
\description{
Evaluate column index of a division object that respect bounding restriction on their corresponding critical values.
}
\examples{
#General example
x <- c(9,11, sample(12:19, 3), 20,22)
division <- group_division(x, n_group = 3, min_group = 1)
group_division_bound(x, division, min_crit = c(11,11), max_crit = c(20,20))

}
