% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_division_check.R
\name{group_division_check}
\alias{group_division_check}
\title{Group division check}
\usage{
group_division_check(n_row, division)
}
\arguments{
\item{n_row}{\code{integer}, the number of row of the matrix to build, which should equal the length of the vector to be classified.}

\item{division}{\code{matrix} of position that define a new group (rows) for different partitions (columns), as generated by the function group_division; alternatively a \code{numeric} vector representing a single partition.}
}
\value{
Return \code{TRUE} if properties of \code{division} are appropriate and an error otherwise.
}
\description{
Check that a division vector or matrix is appropriate.
}
\examples{
#General example
##Data
x <- sample(1:100, 10)
division <- group_division(x, n_group = 2, min_group = 2)

##Case that return TRUE, select n_row = 5 for an example that return an error.
group_division_check(n_row = length(x), division)
group_division_check(n_row = length(x), division[,1])

}
