% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_matrix.R
\name{group_matrix}
\alias{group_matrix}
\title{Group matrix}
\usage{
group_matrix(n_row, division)
}
\arguments{
\item{n_row}{\code{integer}, the number of row of the matrix to build, which should equal the length of the vector to be classified.}

\item{division}{\code{matrix} of position that define a new group (rows) for different partitions (columns), as generated by the function \code{group_division}.}
}
\value{
Return a \code{matrix} of \code{integer} representing the group to which elements of a vector might belong.
}
\description{
Generate a matrix of group, from a division matrix.
}
\details{
Division is checked for validity before generating matrix.
}
\examples{
#General example
x <- sample(1:100, 10)
division <- group_division(x, n_group = 2, min_group = 2)
group_matrix(n_row = length(x), division)

}
