% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nb.R
\name{group_nb}
\alias{group_nb}
\title{Number of group}
\usage{
group_nb(group)
}
\arguments{
\item{group}{\code{matrix} of \code{integer} indicating the group elements belong (rows) for various partition (columns), as generated by the function \code{group_matrix}.}
}
\description{
Provides the number of group in a group matrix.
}
\examples{
group <- t(gtools::permutations(3,3,1:3))
group_nb(group)
}
