% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_selection_matrices.R
\name{group_selection_matrices}
\alias{group_selection_matrices}
\title{Selection matrices}
\usage{
group_selection_matrices(group)
}
\arguments{
\item{group}{\code{matrix} of \code{integer} indicating the group elements belong (rows) for various partition (columns), as generated by the function \code{group_matrix}.}
}
\value{
Return a list of matrix which elements, named after the group, represent if the element belong to the group (1) or not (0).
}
\description{
Generate a list of selection matrices from a group matrix.
}
\details{
The groups in the \code{group} matrix must be integers, incremented from 1.
}
\examples{
group <- t(gtools::permutations(3,3,1:3))
group_selection_matrices(group)

}
