% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_vector.R
\name{group_vector}
\alias{group_vector}
\title{Group matrix}
\usage{
group_vector(n_row, division_vector)
}
\arguments{
\item{n_row}{\code{integer}, the number of row of the matrix to build, which should equal the length of the vector to be classified.}

\item{division_vector}{\code{integer}, vector of position that define a new group (row).}
}
\value{
Return a vector of \code{integer} representing the group to which elements of a vector might belong.
}
\description{
Generate a vector of group, from a division vector.
}
\details{
Division is checked for validity before generating the vector.
}
\examples{
#General example
x <- sample(1:100, 10)
division <- group_division(x, n_group = 2, min_group = 2)
group_vector(n_row = length(x), division[,1])
}
