% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_contingency.R
\name{quadrat_contingency}
\alias{quadrat_contingency}
\title{Quadrat contingency}
\usage{
quadrat_contingency(Q)
}
\arguments{
\item{Q}{\code{numeric} vector corresponding to counts of quadrats.}
}
\value{
Return the corresponding contingency \code{matrix}.
}
\description{
Form a contingency table (\code{matrix}) from quadrat counts.
}
\details{
The order of elements in \code{Q} are assumed to increment by \code{i} first, then \code{j} (e.g. 11, 12, 12, 21, 22, ...), with i designating quadrat indexes along the the x-axis and j the y-axis. As a supplementary constraint, i and j must range over the same values and therefore Q must have a square number of elements (quadrats).
}
\examples{
Q <- rnorm(9)
quadrat_contingency(Q)
}
