% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_count.R
\name{quadrat_count}
\alias{quadrat_count}
\title{Quadrat count}
\usage{
quadrat_count(group_x, group_y)
}
\arguments{
\item{group_x}{group \code{matrix} containing only one partition (column).}

\item{group_y}{group \code{matrix} containing one or multiple partitions (column).}
}
\value{
Return a \code{data.frame} with columns named \code{ij}, after the quadrat, with i the quadrat index on the x-axis and j the quadrat index on the y-axis. Each line correspond to a partition of \code{group_y}.
}
\description{
Count the number of points that is present in quadrat defined by \code{group_x} in x and \code{group_y} in y.
}
\examples{
#Generate data
##Data
n = 10
x <- rnorm(n)
y <- rnorm(n)
##Group x, only one partition
division_x  <- group_division(y, 2, min_group = 2)
group_x <- group_matrix(n, division_x[,1 , drop = FALSE])

##Group y, multiple partition
division  <- group_division(y, 2, min_group = 2)
group_y <- group_matrix(n, division)

#Compute the number of elements in each quadrat
quadrat_count(group_x, group_y)

}
