% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_count_pred.R
\name{quadrat_count_pred}
\alias{quadrat_count_pred}
\title{Quadrat count prediction}
\usage{
quadrat_count_pred(Q, trend)
}
\arguments{
\item{Q}{\code{data.frame} of quadrat counts, as generated by the function \code{quadrat_count}. Must possesses a squared number of elements.}

\item{trend}{\code{character}, either \code{positive} or \code{negative}.}
}
\value{
Return a \code{data.frame} with columns named \code{ij}, after the quadrat, with i the quadrat index on the x-axis and j the quadrat index on the y-axis. Each line correspond to a partition of \code{group_y}.
}
\description{
Count the number of points in quadrats on the diagonal specified by \code{trend}, from existing an data count.
}
\examples{
#Generate data
##Data
n = 10
x <- rnorm(n)
y <- rnorm(n)
##Group x, only one partition
division_x  <- group_division(y, 2, min_group = 2)
group_x <- group_matrix(n, division_x[,1 , drop = FALSE])

##Group y, multiple partition
division  <- group_division(y, 2, min_group = 2)
group_y <- group_matrix(n, division)

#Compute the number of elements in each quadrat
Q <- quadrat_count(group_x, group_y)

#Compute the number of element on the positive diagonal
quadrat_count_pred(Q = Q, trend = "positive")

}
