% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_name.R
\name{quadrat_name}
\alias{quadrat_name}
\title{Quadrat name}
\usage{
quadrat_name(n_group)
}
\arguments{
\item{n_group}{\code{integer}, the number of groups.}
}
\value{
Return a vector of name for the quadrats (\code{character}).
}
\description{
Generate names of quadrats.
}
\details{
The order of elements are incremented by \code{i} first, then \code{j} (e.g. 11, 12, 12, 21, 22, ...), with i designating quadrat indexes along the the x-axis and j the y-axis. As a supplementary constraint, i and j range from 1 to the \code{n_group}, representing a squared number of quadrats. Names are non ambiguous for n_group smaller or equal to 10
}
\examples{
n_group = 10
quadrat_name(n_group)
}
