% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_name_err.R
\name{quadrat_name_err}
\alias{quadrat_name_err}
\title{Quadrat name error}
\usage{
quadrat_name_err(n_group, trend)
}
\arguments{
\item{n_group}{\code{integer}, the number of groups.}

\item{trend}{\code{character}, either \code{positive} or \code{negative}.}
}
\value{
Return a vector of quadrat names.
}
\description{
Generate name of quadrats that does not correspond to the diagonal specified by \code{trend}. Complement of the function \code{quadrat_name_pred}.
}
\examples{
quadrat_name_pred(n_group = 3, trend = "positive")
quadrat_name_err(n_group = 3, trend = "positive")
}
