% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_name_pred.R
\name{quadrat_name_pred}
\alias{quadrat_name_pred}
\title{Quadrat name prediction}
\usage{
quadrat_name_pred(n_group, trend)
}
\arguments{
\item{n_group}{\code{integer}, the number of groups.}

\item{trend}{\code{character}, either \code{positive} or \code{negative}.}
}
\value{
Return a vector of quadrat names.
}
\description{
Generate name of quadrats that correspond to the diagonal specified by \code{trend}.
}
\examples{
quadrat_name_pred(n_group = 3, trend = "positive")
quadrat_name_pred(n_group = 3, trend = "negative")
}
