% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_stat.R
\name{quadrat_stat}
\alias{quadrat_stat}
\title{Quadrat statistics}
\usage{
quadrat_stat(Q)
}
\arguments{
\item{Q}{\code{numeric} vector corresponding to counts of quadrats.}
}
\description{
Compute the Cramer's V and Fisher p values of the contingency matrix made by the quadrats.
}
\details{
The order of elements in \code{Q} are assumed to increment by \code{i} first, then \code{j} (e.g. 11, 12, 12, 21, 22, ...), with i designating quadrat indexes along the the x-axis and j the y-axis. As a supplementary constraint, i and j must range over the same values and therefore Q must have a square number of elements (quadrats).
}
\examples{
#Generate data
##Data
n = 10
x <- rnorm(n)
y <- rnorm(n)
##Group x, only one partition
division_x  <- group_division(y, 2, min_group = 2)
group_x <- group_matrix(n, division_x[,1 , drop = FALSE])

##Group y, multiple partition
division  <- group_division(y, 2, min_group = 2)
group_y <- group_matrix(n, division)

#Compute the number of elements in each quadrat
Q <- quadrat_count(group_x, group_y)
quadrat_stat(Q)
}
