% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_division.R
\name{group_division}
\alias{group_division}
\title{Group division}
\usage{
group_division(x, n_group, min_group, min_crit = NULL, max_crit = NULL)
}
\arguments{
\item{x}{\code{numeric}, an ordered vector of elements.}

\item{n_group}{\code{integer}, the number of groups (minimum 2).}

\item{min_group}{\code{integer}, the minimum number of values in each group (not elements).}

\item{min_crit, max_crit}{\code{numeric} vectors of length corresponding to the number of group (number of rows of \code{division}) minus one, representing inequality constraints on each critical values in \code{x} \code{(min <= x <= max)}. The default value is \code{NULL} and represent no constraint.}
}
\value{
Return a \code{matrix} of position that define new groups (divisions). Rows represent the \code{n_group - 1} divisions for a given partition, while columns represent different partitions.
}
\description{
Find the possible ways to separate an ordered numerical vector in groups.
}
\details{
Repeated values are kept in the same group and only count as one value with respect to the \code{min_group} constraint. An error is thrown if no group division meet the conditions imposed.
}
\examples{
#General example
x <- sample(1:100, 10)
group_division(x, n_group = 2, min_group = 2)

#Example with a repeated value
x <- 100*c(1:3, 3, 3:5)
group_division(x, n_group = 2, min_group = 2)

}
